requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
pandas==2.0.3
openpyxl==3.1.2
dnspython==2.4.2
eventlet==0.33.3
colorama==0.4.6
python-socketio==5.8.0